# core-lending-mortgages-sdk/target/classes/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-mortgages
  description: Mortgages Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.catalis.vc/mortgages
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/mortgage-contracts/{mortgageContractId}/payment-records/{paymentRecordId}:
    get:
      tags:
        - MortgagePaymentRecord
      summary: Get a payment record by ID
      operationId: getById_3
      parameters:
        - name: mortgageContractId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: paymentRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MortgagePaymentRecordDTO'
    put:
      tags:
        - MortgagePaymentRecord
      summary: Update a payment record
      operationId: update_3
      parameters:
        - name: mortgageContractId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: paymentRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MortgagePaymentRecordDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MortgagePaymentRecordDTO'
    delete:
      tags:
        - MortgagePaymentRecord
      summary: Delete a payment record
      operationId: delete_3
      parameters:
        - name: mortgageContractId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: paymentRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/mortgage-contracts/{mortgageContractId}/payment-records:
    get:
      tags:
        - MortgagePaymentRecord
      summary: List or search mortgage payment records for a contract
      operationId: findAll_3
      parameters:
        - name: mortgageContractId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.mortgageContractId
          in: query
          description: Exact filter for mortgage contract id
          required: false
          schema:
            type: string
        - name: filters.paymentScheduleId
          in: query
          description: Exact filter for payment schedule id
          required: false
          schema:
            type: string
        - name: filters.transactionId
          in: query
          description: Exact filter for transaction id
          required: false
          schema:
            type: string
        - name: filters.paymentAmount
          in: query
          description: Exact filter for payment amount
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[paymentAmount].from
          in: query
          description: Filter payment amount from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[paymentAmount].to
          in: query
          description: Filter payment amount to value
          required: false
          schema:
            type: string
        - name: filters.paymentDate
          in: query
          description: Exact filter for payment date
          required: false
          schema:
            type: string
        - name: filters.isPartialPayment
          in: query
          description: Exact filter for is partial payment
          required: false
          schema:
            type: boolean
        - name: filters.note
          in: query
          description: Exact filter for note
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseMortgagePaymentRecordDTO'
    post:
      tags:
        - MortgagePaymentRecord
      summary: Create a new payment record
      operationId: create_3
      parameters:
        - name: mortgageContractId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MortgagePaymentRecordDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MortgagePaymentRecordDTO'
...
components:
  schemas:
...
    MortgagePaymentRecordDTO:
      required:
        - isPartialPayment
        - mortgageContractId
        - paymentAmount
        - paymentDate
        - paymentScheduleId
        - transactionId
      type: object
      properties:
        paymentRecordId:
          type: string
          format: uuid
          readOnly: true
        mortgageContractId:
          type: string
          format: uuid
        paymentScheduleId:
          type: string
          format: uuid
        transactionId:
          type: string
          format: uuid
        paymentAmount:
          maximum: 999999.99
          exclusiveMaximum: false
          minimum: 0.01
          exclusiveMinimum: false
          type: number
        paymentDate:
          type: string
          format: date
        isPartialPayment:
          type: boolean
        note:
          maxLength: 500
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationResponseMortgagePaymentRecordDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/MortgagePaymentRecordDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...